Revert "将AD值转化为0-100直接传上去"
This reverts commit 3dd801613f164f1300c02b538107c25e89875b74.

+++ MyCode/can/can.h
@@ -7,7 +7,7 @@
 										 							 				    
 u8 CAN_Mode_Init(u8 tsjw,u8 tbs2,u8 tbs1,u16 brp,u8 mode);//CAN初始化
  
-u8 Can_Send_Msg(int* msg,u8 len);						//发送数据
+u8 Can_Send_Msg(u8* msg,u8 len);						//发送数据
 
 u8 Can_Receive_Msg(u8 *buf);							//接收数据
 
@@ -21,5 +21,5 @@ typedef struct
 	u8 guzhang_flag;
 }deflag;
 extern deflag flag;
-extern int cansend[4];
+extern u8 cansend[4];
 #endif
